+++ src/components/layout/SideNav/SideNav.tsx
@@ -18,7 +18,7 @@ import { UserProductRole } from '@/types/party.types'
 import { useJwt } from '@/hooks/useJwt'
 import { SELFCARE_BASE_URL } from '@/config/env'
 import { useCurrentRoute } from '@/router'
-import { getParentRoutes } from '@/router/utils'
+import { getParentRoutes } from '@/router/router.utils'
 import { SIDENAV_WIDTH } from '@/config/constants'
 import { SideNavItemLink, SideNavItemLinkSkeleton } from './SideNavItemLink'
 import { CollapsableSideNavItem, CollapsableSideNavItemSkeleton } from './CollapsableSideNavItem'
