+++ src/pages/ProviderEServiceCreatePage/components/EServiceCreateStep1General/EServiceCreateStep1General.tsx
@@ -14,7 +14,7 @@ import { EServiceMutations } from '@/api/eservice'
 import { FrontendAttributes } from '@/types/attribute.types'
 import { remapEServiceAttributes } from '@/utils/attribute.utils'
 import { remapFrontendAttributesToBackend } from '@/api/eservice/eservice.api.utils'
-import { URL_FRAGMENTS } from '@/router/utils'
+import { URL_FRAGMENTS } from '@/router/router.utils'
 import useCurrentLanguage from '@/hooks/useCurrentLanguage'
 import { useJwt } from '@/hooks/useJwt'
 import { getKeys } from '@/utils/array.utils'
