+++ src/pages/ProviderEServiceListPage/components/EServiceTableRow.tsx
@@ -4,7 +4,7 @@ import { StatusChip, StatusChipSkeleton } from '@/components/shared/StatusChip'
 import { Box, Button, Skeleton, Stack } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 import { useNavigateRouter } from '@/router'
-import { URL_FRAGMENTS } from '@/router/utils'
+import { URL_FRAGMENTS } from '@/router/router.utils'
 import useCurrentLanguage from '@/hooks/useCurrentLanguage'
 import { ActionMenu, ActionMenuSkeleton } from '@/components/shared/ActionMenu'
 import { EServiceProvider } from '@/types/eservice.types'
