+++ src/router/hooks/useNavigateRouter.ts
@@ -1,5 +1,5 @@
 import { GetRouteUrl, Navigate } from '@/router/router.types'
-import { getLocalizedRoutePathname } from '@/router/utils'
+import { getLocalizedRoutePathname } from '@/router/router.utils'
 import { useCallback } from 'react'
 import { generatePath, useNavigate } from 'react-router-dom'
 import useCurrentLanguage from '@/hooks/useCurrentLanguage'
