+++ src/router/hooks/useRouteParams.ts
@@ -2,7 +2,7 @@ import { NotFoundError } from '@/utils/errors.utils'
 import { useParams } from 'react-router-dom'
 import { RouteKey, RouteParams } from '../router.types'
 import useCurrentRoute from './useCurrentRoute'
-import { getDynamicPathsName } from '../utils'
+import { getDynamicPathSegments } from '../router.utils'
 
 /**
  * Wrapper for react-router-dom's useParams that performs a runtime check on the actual route params.
@@ -11,7 +11,7 @@ function useRouteParams<T extends RouteKey>(): RouteParams<T> {
   const { routeKey } = useCurrentRoute()
   const params = useParams()
 
-  const dynamicPathNames = getDynamicPathsName(routeKey)
+  const dynamicPathNames = getDynamicPathSegments(routeKey)
 
   if (!dynamicPathNames.every((dynamicPathName) => dynamicPathName in params)) {
     throw new NotFoundError()
